requirements.txt
amqp==5.1.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
billiard==4.1.0
botocore==1.35.36
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.3
fastapi==0.115.2
fire==0.5.0
greenlet==2.0.2
h11==0.14.0
pydantic==2.9.2
jmespath==1.0.1
kombu==5.3.1
numpy==1.25.2
prompt-toolkit==3.0.39
python-dateutil>=2.8.2
python-dotenv==1.0.0
redis==5.0.0
s3transfer
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
termcolor==2.3.0
typing_extensions
tzdata==2023.3
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
httpx~=0.27.2
langchain-openai
starlette~=0.40.0
requests~=2.32.3
langchain~=0.3.7
PyMySQL==1.1.1
cryptography==41.0.3
SQLAlchemy==2.0.20
langchain-community==0.3.3
asyncpg==0.30.0
openai~=1.51.2
pyjwt~=2.9.0
nanoid==2.0.0
pandas==2.2.1
rapidfuzz~=3.5.2
python-multipart==0.0.6
numpy_financial==1.0.0
google-cloud-pubsub==2.27.1
firebase-admin==6.5.0
langchain-aws
boto3==1.35.36
python-magic==0.4.27
aiohttp~=3.10.10
pillow>=11.1.0
joblib~=1.4.2
anthropic-bedrock==0.8.0
PyMuPDF==1.24.14
celery
alembic==1.14.0
apscheduler==3.11.0
feedparser==6.0.11
aioboto3==13.2.0
beautifulsoup4==4.12.3
validators===0.34.0
langchain-google-genai===2.0.11
google-api-python-client===2.162.0
google-auth-oauthlib===1.2.1
google-auth-httplib2===0.2.0
isodate===0.7.2
youtube-transcript-api===0.6.3
markitdown[all]==0.1.0a1
python-docx==1.1.2
markdown==3.7
fal-client==0.5.9
mutagen==1.47.0
langcodes==3.5.0
ffmpeg-python==0.2.0
ragflow-sdk===0.18.0
langdetect===1.0.9
langid===1.1.6